Update custom errors for user pipeline

+++ src/repository/userRepository.ts
@@ -1,12 +1,9 @@
 import { prisma } from "@/lib/prisma/prisma";
 import { User } from "../app/domain/entities/user";
 import IUserRepository from "@/app/domain/repositories/IUserRepository";
-import { getLogger } from "@/lib/log/logUtil";
-
-const logger = getLogger("account");
 
 export class UserRepository implements IUserRepository {
-  async createUser(username: string, email: string, pwHashed: string): Promise<User> {
+  async createUser(username: string, email: string, pwHashed: string): Promise<User | undefined> {
     const user = await prisma.users.create({
       data: {
         email: email,
@@ -16,8 +13,7 @@ export class UserRepository implements IUserRepository {
     });
 
     if (!user) {
-      logger.error(`[userRepository] error creating account.`); 
-      throw new Error("[userRepository] error creating account.");
+      return undefined;
     }
   
     return user
@@ -27,7 +23,6 @@ export class UserRepository implements IUserRepository {
     const user = await prisma.users.findUnique({ where: { email: email }});
 
     if (!user) {
-      logger.info(`[userRepository] email ${email} does not exist.`); 
       return undefined;
     }
 
@@ -38,7 +33,6 @@ export class UserRepository implements IUserRepository {
     const user = await prisma.users.findUnique({ where: { username: username }});
 
     if (!user) {
-      logger.info(`[userRepository] username ${username} does not exist.`); 
       return undefined;
     }
 
@@ -46,6 +40,11 @@ export class UserRepository implements IUserRepository {
   }
 
   async updateUserPassword(email: string, newHashedPassword: string): Promise<User | undefined> {
+    const existingUser = await this.getUserByEmail(email);
+    if (!existingUser) {
+      return undefined;
+    }
+
     const user = await prisma.users.update({
       where: {
         email: email,
@@ -55,25 +54,20 @@ export class UserRepository implements IUserRepository {
       }
     });
 
-    if (!user) {
-      logger.error(`[userRepository] user with ${email}, failed to update password.`); 
-      return undefined;
-    }
-
     return user;
   }
 
   async deleteUserByEmail(email: string): Promise<User | undefined> {
+    const existingUser = await this.getUserByEmail(email);
+    if (!existingUser) {
+      return undefined;
+    }
+
     const deletedUser = await prisma.users.delete({
       where: {
         email: email,
       },
     });
-
-    if (!deletedUser) {
-      logger.error(`[userRepository] user with ${email}, failed to delete account.`); 
-      return undefined;
-    }
 
     return deletedUser;
   }
